third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-cascade-trick.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e1 {
    --props: {
        width: 100px;
        background-color: red;
    }
}
#e2 {
    --props-copy: var(--props);
}
#e3 {
    --props: {
        @apply --props-copy;
        background-color: green;
        height: 200px;
    }
}
#e4 {
    @apply --props;
}
</style>
<div id=e1>
  <div id=e2>
    <div id=e3>
      <div id=e4>
      </div>
    </div>
  </div>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e4).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e4).width, "100px");
    assert_equals(getComputedStyle(e4).height, "200px");
}, "Custom property sets can be extended with the cascade trick");
</script>
